PvP revisions. Fixed bug where battle urgency (and features like centroid) were always 0.

+++ src/ProxyBwapi/UnitInfo/UnitInfo.scala
@@ -221,7 +221,7 @@ abstract class UnitInfo(baseUnit: bwapi.Unit, id: Int) extends UnitProxy(baseUni
         (is(Zerg.Ultralisk)   && player.hasUpgrade(Zerg.UltraliskSpeed)))
         1.5 else 1.0)))
   
-  def projectFrames(framesToLookAhead: Int): Pixel = pixelCenter.add((velocityX * framesToLookAhead).toInt, (velocityY * framesToLookAhead).toInt)
+  def projectFrames(framesToLookAhead: Double): Pixel = pixelCenter.add((velocityX * framesToLookAhead).toInt, (velocityY * framesToLookAhead).toInt)
   
   def inTileRadius  (tiles: Int)  : Traversable[UnitInfo] = With.units.inTileRadius(tileIncludingCenter, tiles)
   def inPixelRadius (pixels: Int) : Traversable[UnitInfo] = With.units.inPixelRadius(pixelCenter, pixels)
